backend/requirements.txt
# Localix Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
ollama==0.1.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
apscheduler==3.10.4
networkx==3.2.1
jinja2==3.1.3
markdown==3.5.2
weasyprint==60.2
